api_backend/requirements.txt
# Core FastAPI
fastapi>=0.115.2
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart>=0.0.18

# Database & Storage
supabase==2.9.0

# Environment & Configuration
python-dotenv==1.0.1

# Analytics & Parsing
user-agents==2.2.0
httpx==0.27.2

# Monitoring & Logging
structlog==24.4.0
prometheus-client==0.21.0

# Utilities
python-json-logger==2.0.7
